requirements.txt
# TradingAgents Requirements

# LangChain & LLM
typing-extensions
langchain-openai
langchain-experimental
langchain_anthropic
langchain-google-genai
langgraph
chromadb

# Data Analysis
pandas
numpy
yfinance
stockstats
backtrader

# Data Sources
finnhub-python
eodhd
akshare
tushare

# Web Scraping
requests
beautifulsoup4
lxml
parsel

# Crypto
python-binance

# News & Social
praw
feedparser

# Dashboard
streamlit>=1.31.0
plotly>=5.18.0

# Database & Utils
redis
cryptography>=41.0.0
pytz
tqdm
setuptools

# CLI
chainlit
rich
questionary
